+++ assets/js/admin.js
@@ -1,31 +1,1 @@
-jQuery(document).ready(function($) {
-
-    let degreeSelect = $('select#fau-degree-programm-shares-degree-select');
-    let subjectSelect = $('select#fau-degree-programm-shares-subject-select');
-    let shortcodeTemplate = $('code.fau-degree-programm-shares-shortcode-template');
-
-    degreeSelect.change(function() {
-        setShortcodeTemplate();
-    });
-
-    subjectSelect.change(function() {
-        setShortcodeTemplate();
-    });
-
-    function setShortcodeTemplate() {
-        let degreeID = degreeSelect.val() ;
-        if (degreeID === '0') degreeID = '';
-        let subjectID = subjectSelect.val();
-        if (subjectID === '0') subjectID = '';
-
-        $('#fau-degree-program-shares span.dashicons.dashicons-yes').remove();
-        shortcodeTemplate.html('[fachanteile abschluss="' + degreeID + '" fach="' + subjectID + '"]');
-    }
-
-    $('button.fau-degree-programm-shares-copy-shortcode').click(function() {
-        let copyText = shortcodeTemplate.html();
-        navigator.clipboard.writeText(copyText);
-        $(this).parent().find('span.dashicons.dashicons-yes').remove();
-        $(this).after('<span class="dashicons dashicons-yes" style="font-size: 2em; color: #2271b1"></span>');
-    });
-});
+jQuery(document).ready(function($){let degreeSelect=$("select#fau-degree-programm-shares-degree-select");let subjectSelect=$("select#fau-degree-programm-shares-subject-select");let percentCheck=$("input#fau-degree-programm-shares-percent-check");let shortcodeTemplate=$("code.fau-degree-programm-shares-shortcode-template");degreeSelect.change(function(){setShortcodeTemplate()});subjectSelect.change(function(){setShortcodeTemplate()});percentCheck.change(function(){setShortcodeTemplate()});function setShortcodeTemplate(){let degreeID=degreeSelect.val();if(degreeID==="0")degreeID="";let subjectID=subjectSelect.val();if(subjectID==="0")subjectID="";let showPercent=percentCheck.is(":checked");let textPercent="";if(showPercent===false){textPercent=' percent="0"'}$("#fau-degree-program-shares span.dashicons.dashicons-yes").remove();shortcodeTemplate.html('[fachanteile abschluss="'+degreeID+'" fach="'+subjectID+'"'+textPercent+"]")}$("button.fau-degree-programm-shares-copy-shortcode").click(function(){let copyText=shortcodeTemplate.html();navigator.clipboard.writeText(copyText);$(this).parent().find("span.dashicons.dashicons-yes").remove();$(this).after('<span class="dashicons dashicons-yes" style="font-size: 2em; color: #2271b1"></span>')})});
