+++ src/scss/admin.scss
@@ -1,6 +1,6 @@
 .wp-admin {
     #fau-degree-program-shares {
-        label {
+        label.block-label {
             display: block;
             padding: .2em 0;
         }
